+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 ### Fixed
 - Fixed an error that occurred when running CLI commands on Windows. ([#8089](https://github.com/craftcms/cms/issues/8089))
+- Fixed an error that could occur when updating to Craft 3.7 if there were any soft-deleted global sets. ([#8079](https://github.com/craftcms/cms/issues/8079))
 
 ## 3.7.0-beta.2 - 2021-06-17